fix: `u32` was defined mistakenly as `uint8_t`

+++ tools/skap/include/skap_defines.h
@@ -29,6 +29,6 @@
 #define skap_return_defer(x) do { result = 1; goto defer; } while(0)
 
 typedef uint8_t u8;
-typedef uint8_t u32;
+typedef uint32_t u32;
 typedef uint64_t u64;
 typedef size_t usz;
